requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.5
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
websockets==13.1
grpcio==1.68.0
grpcio-tools==1.68.0
openai==1.54.0
supabase==2.28.3
sqlalchemy==2.0.36
asyncpg==0.30.0
redis==5.2.0
python-dotenv==1.0.1
httpx==0.27.2
numpy==2.1.3
scikit-learn==1.5.2
email-validator==2.3.0

